feat: consolidate vesting tests
